JournalPeriod: [wip] subsume BeankeepPeriod
